Refactored #to_csv into concern for code climate
